+++ live-examples/interop/method-discovery-by-event/app-b/index.js
@@ -1,0 +1,37 @@
+/* eslint-disable no-undef */
+const APP_NAME = 'App B';
+
+// Entry point. Initializes Glue42 Web. А Glue42 Web instance will be attached to the global window.
+window.startApp({ appName: APP_NAME })
+  .then(() => {
+    document.getElementById("registerGlueMethodBtn")
+      .addEventListener('click', registerGlueMethodHandler);
+  })
+  .then(clearLogsHandler)
+  .catch(console.error);
+
+function registerGlueMethodHandler() {
+  const methodNameInput = document.getElementById('methodNameInput');
+  registerGlueMethod(methodNameInput.value);
+  methodNameInput.value = '';
+}
+
+async function registerGlueMethod(methodName) {
+  const methodDefinition = { name: methodName };
+  const invocationHandler = () => {
+    logger.info(`Method "${methodDefinition.name}" invoked.`);
+
+    return {
+      result: `Hello from "${methodDefinition.name}" in "${APP_NAME}"!`
+    };
+  }
+
+  try {
+    await glue.interop.register(methodDefinition, invocationHandler)
+
+    logger.info(`Method "${methodDefinition.name}" registered.`);
+  } catch (error) {
+    console.error(`Failed to register "${methodDefinition.name}". Error: `, error);
+    logger.error(`Failed to register "${methodDefinition.name}".`);
+  }
+}