+++ live-examples/interop/stream-subscription-request/app-a/index.js
@@ -1,0 +1,7 @@
+/* eslint-disable no-undef */
+const APP_NAME = 'App A';
+
+// Entry point. Initializes Glue42 Web. А Glue42 Web instance will be attached to the global window.
+window.startApp({ appName: APP_NAME })
+  .then(() => window.onAppStarted())
+  .catch(console.error);