+++ packages/core/src/interop/client/repository.ts
@@ -78,6 +78,7 @@ export default class ClientRepository {
         }
 
         const identifier = this.createMethodIdentifier(method);
+        const that = this;
         const methodDefinition: ClientMethodInfo = {
             identifier,
             gatewayId: method.id,
@@ -89,16 +90,14 @@ export default class ClientRepository {
             accepts: method.input_signature,
             returns: method.result_signature,
             supportsStreaming: typeof method.flags !== "undefined" ? method.flags.streaming : false,
-
+            getServers: () => {
+                return that.getServersByMethod(identifier);
+            }
         };
         // now add some legacy stuff
         (methodDefinition as any).object_types = methodDefinition.objectTypes;
         (methodDefinition as any).display_name = methodDefinition.displayName;
         (methodDefinition as any).version = methodDefinition.version;
-        const that = this;
-        methodDefinition.getServers = () => {
-            return that.getServersByMethod(method.id);
-        };
 
         server.methods[method.id] = methodDefinition;
 
@@ -178,7 +177,7 @@ export default class ClientRepository {
     public onMethodAdded(callback: (method: ClientMethodInfo) => void): UnsubscribeFunction {
         const unsubscribeFunc = this.callbacks.add("onMethodAdded", callback);
 
-        // because we need the servers shapshot before we return to the application code
+        // because we need the servers snapshot before we return to the application code
         const methodsToReplay = this.getMethods();
 
         return this.returnUnsubWithDelayedReplay(unsubscribeFunc, methodsToReplay, callback);
@@ -249,12 +248,13 @@ export default class ClientRepository {
         return (methodInfo.name + accepts + returns).toLowerCase();
     }
 
-    private getServersByMethod(id: string): Glue42Core.AGM.Instance[] {
+    private getServersByMethod(identifier: string): Glue42Core.AGM.Instance[] {
         const allServers: Glue42Core.AGM.Instance[] = [];
         Object.keys(this.servers).forEach((serverId) => {
             const server = this.servers[serverId];
             Object.keys(server.methods).forEach((methodId) => {
-                if (methodId === id) {
+                const methodInfo = server.methods[methodId];
+                if (methodInfo.identifier === identifier) {
                     allServers.push(server.instance);
                 }
             });
@@ -262,7 +262,7 @@ export default class ClientRepository {
         return allServers;
     }
 
-    // collectionToReplay: because we need a shapshot before we exist this stack
+    // collectionToReplay: because we need a snapshot before we exist this stack
     private returnUnsubWithDelayedReplay(unsubscribeFunc: UnsubscribeFunction, collectionToReplay: any[], callback: any) {
 
         // because we want to interrupt the loop with the existing methods