+++ tutorials/vanilla-js/start/Stocks/index.js
@@ -131,7 +131,6 @@ const newPricesHandler = (priceUpdate) => {
 };
 
 const stockClickedHandler = (stock) => {
-    console.log(stock);
     sessionStorage.setItem('stock', JSON.stringify(stock));
     window.location.href = `http://${window.location.host}/stocks/details/index.html`;
 };